Update dashboard.py
Data for dashboard directory change

rev_day_dir and rev_hour_dir directory changed as this wasn't fit with the README.txt context caused directory not found. 

- Previously: ../"dashboard"/"rev_day.csv" ;
- Now: ../"rev_day.csv"

This is required because in README.txt current cwd for streamlit run is already on dashboard folder. and the data is already there.